restructured radioGroup folder
