+++ src/components/radioGroup/radioGroup/index.ts
@@ -1,0 +1,1 @@
+export { RadioGroup, type IRadioGroupProps } from './radioGroup';
